requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-translate==3.1.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pdfminer==20191125
proto-plus==1.18.1
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
